Codes
These classes are related to the traveling salesman game project.  I have written this game in Java language and I have used audio and video effects and the swing graphic library, and I have written the music using the Jay Lear library.  In this game, I provided the user with interesting features such as playing the desired music, choosing the character and choosing the theme of the game.  And this is how we challenge you.  I put these codes for you, if you have any questions about it, I will respond.  Thank you